pyproject.toml
[project]
name = "mcps"
version = "0.1.0"
description = "Model Context Protocol System - Knowledge Graph for the MCP Ecosystem"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Database & ORM
    "sqlmodel==0.0.16",  # Pydantic v2 compatible
    "alembic>=1.17.2",
    "sqlalchemy>=2.0.44",

    # PostgreSQL drivers (production)
    "asyncpg>=0.29.0",  # Async PostgreSQL driver (fastest)
    "psycopg2-binary>=2.9.9",  # Sync driver for migrations and pgAdmin

    # SQLite drivers (legacy/dev only)
    "aiosqlite>=0.21.0",  # Optional: for backward compatibility

    # Vector Search
    "pgvector>=0.2.4",  # PostgreSQL vector extension
    "sqlite-vec>=0.1.0",  # Optional: for SQLite backward compatibility

    # Settings & Config
    "pydantic-settings>=2.12.0",

    # HTTP & Network
    "httpx>=0.27.0",
    "tenacity>=9.0.0",

    # Logging & CLI
    "loguru>=0.7.3",
    "typer>=0.15.0",

    # Data Export
    "pyarrow>=18.0.0",

    # Parsing & Analysis
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-typescript>=0.23.0",

    # Package Inspection
    "pkginfo>=1.11.0",

    # API Framework (for apps/api)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Social Media APIs
    "praw>=7.8.1",  # Reddit API (synchronous, we'll wrap in asyncio)
    "tweepy>=4.15.0",  # Twitter/X API
    "google-api-python-client>=2.150.0",  # YouTube Data API
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",

    # NLP & Sentiment Analysis
    "vaderSentiment>=3.3.2",  # Sentiment analysis

    # Caching & Redis
    "redis>=5.0.0",  # Redis client
    "redis-om>=0.2.0",  # Redis object mapper

    # Metrics & Monitoring
    "prometheus-client>=0.19.0",  # Prometheus metrics

    # Error Tracking
    "sentry-sdk[fastapi]>=1.39.0",  # Sentry integration (optional)

    # Rate Limiting
    "slowapi>=0.1.9",  # Rate limiting for FastAPI

    # Task Scheduling
    "apscheduler>=3.10.0",  # Background task scheduling

    # Supabase Integration
    "supabase>=2.9.0",  # Official Supabase Python client
    "gotrue>=2.8.0",  # Supabase Auth
    "postgrest-py>=0.16.0",  # Supabase PostgREST client
    "storage3>=0.8.0",  # Supabase Storage
    "realtime>=2.0.0",  # Supabase Realtime
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-pyyaml>=6.0.0",
]
docs = [
    "sphinx>=7.0.0",
    "shibuya>=2024.0.0",
    "myst-parser>=2.0.0",
    "sphinx-design>=0.5.0",
    "sphinx-copybutton>=0.5.0",
    "sphinx-autodoc-typehints>=2.0.0",
    "sphinx-tabs>=3.4.0",
    "sphinx-togglebutton>=0.3.0",
    "sphinxemoji>=0.3.0",
    "sphinxcontrib-mermaid>=0.9.0",
    # "autodoc2>=0.5.0",  # Temporarily disabled due to availability issues
    "autoclasstoc>=1.6.0",
    "sphinx-click>=5.0.0",
    "sphinx-jsonschema>=1.19.0",
    "sphinxcontrib-httpdomain>=1.8.0",
    # "sphinx-git>=11.0.0",  # Temporarily disabled
    # "hoverxref>=1.3.0",  # Temporarily disabled
    "sphinx-sitemap>=2.5.0",
    "sphinx-notfound-page>=1.0.0",
]
notebook = [
    "ipykernel>=7.1.0",
    "ipython>=9.7.0",
    "ipywidgets>=8.1.8",
    "jupysql>=0.11.1",
    "jupyter-ai>=2.31.6",
    "jupyter-server>=2.17.0",
    "jupyterlab>=4.4.10",
    "jupyterlab-code-formatter>=3.0.2",
    "jupyterlab-git>=0.51.2",
    "jupyterlab-github>=4.0.0",
    "jupyterlab-lsp>=5.2.0",
    "jupytext>=1.18.1",
    "nbconvert>=7.16.6",
    "nbdime>=4.0.2",
    "notebook>=7.4.7",
    "papermill>=2.6.0",
    "python-lsp-server[all]>=1.13.1",
]
test = [
    "allure-pytest>=2.15.0",
    "hypothesis>=6.148.2",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-benchmark>=5.2.3",
    "pytest-cov>=7.0.0",
    "pytest-emoji>=0.2.0",
    "pytest-html>=4.1.1",
    "pytest-icdiff>=0.9",
    "pytest-instafail>=0.5.0",
    "pytest-mock>=3.15.1",
    "pytest-randomly>=4.0.1",
    "pytest-sugar>=1.1.1",
    "pytest-timeout>=2.4.0",
    "pytest-xdist>=3.8.0",
    "syrupy>=5.0.0",
]
